+++ .devcontainer/dev/run-fio-nfs.sh
@@ -74,7 +74,7 @@
 
 set -eo pipefail
 
-VERSION="1.1.0-beta.2"
+VERSION="1.1.0-beta.3"
 
 # terminal colors
 SHELL_RED='\033[0;31m'