+++ deployment/terraform-module-knfsd/modules/loadbalancer/main.tf
@@ -7,7 +7,7 @@ terraform {
   required_providers {
     aws = {
       source  = "hashicorp/aws"
-      version = "~> 6.59.0"
+      version = "~> 6.63.0"
     }
     dns = {
       source  = "hashicorp/dns"
@@ -16,7 +16,7 @@ terraform {
   }
   provider_meta "aws" {
     user_agent = [
-      "knfsd-file-cache/modules/loadbalancer/1.1.0-beta.2"
+      "knfsd-file-cache/modules/loadbalancer/1.1.0-beta.3"
     ]
   }
 }