+++ image/resources/scripts/20_post_build.sh
@@ -85,7 +85,7 @@ function remove_packages() (
 # install latest ena driver
 function install_ena_driver() (
 	begin_command "installing ENA driver"
-	git_clone --depth 1 --branch ena_linux_2.17.2 https://github.com/amzn/amzn-drivers.git amzn-drivers
+	git_clone --depth 1 --branch ena_linux_2.17.3 https://github.com/amzn/amzn-drivers.git amzn-drivers
 	cd amzn-drivers/kernel/linux/ena/
 	make ${MAKE_VERBOSITY}
 	# ena.ko OR ena.ko.zst