+++ image/smoke-tests/modules/nfs-client/main.tf
@@ -11,12 +11,12 @@ terraform {
   required_providers {
     aws = {
       source  = "hashicorp/aws"
-      version = "~> 6.59.0"
+      version = "~> 6.63.0"
     }
   }
   provider_meta "aws" {
     user_agent = [
-      "knfsd-file-cache/image/smoke-tests/modules/nfs-client/1.1.0-beta.2"
+      "knfsd-file-cache/image/smoke-tests/modules/nfs-client/1.1.0-beta.3"
     ]
   }
 }